Added support for multiple RootNodes
All peers can now start with an array of root nodes instead of a single one,
which removes the SPOF of having a peer depend on a single root node foor joining the cluster.
The method that supports this is GetRootNodes() // notice the plural form //

The GetRandomNode() method returns a random root node, this is for load balancing across VPs.